+++ changelog.rst
@@ -14,6 +14,8 @@ Changelog
 
 **Other Changes**
 
+- [windows] Deleting the observed directory is now detected reliably instead of
+  reporting every ``ReadDirectoryChangesW`` failure as a deletion. (`#1154 <https://github.com/gorakhargosh/watchdog/issues/1154>`__)
 - [windows] Changing only the case of a name no longer emits a deletion event on top of the move event. (`#752 <https://github.com/gorakhargosh/watchdog/issues/752>`__)
 - [bsd] Do not let a ``PermissionError`` raised while registering a kevent kill the emitter thread. (`#1115 <https://github.com/gorakhargosh/watchdog/issues/1115>`__)
 - [core] Add context manager support to ``Observer`` class. The observer can now be used with a ``with`` statement for automatic start/stop management. (`#1090 <https://github.com/gorakhargosh/watchdog/pull/1149>`__)